+++ setuptools_rust/patch.py
@@ -1,3 +1,5 @@
+from distutils.command.install import install
+from distutils.dist import Distribution as DistDistribution
 from setuptools.dist import Distribution
 
 
@@ -18,10 +20,33 @@ def monkey_patch_dist(build_ext):
     Distribution.get_command_class = get_command_class
 
     # use custom has_ext_modules
-    Distribution.orig_has_ext_modules = Distribution.has_ext_modules
+    DistDistribution.orig_has_ext_modules = DistDistribution.has_ext_modules
 
     def has_ext_modules(self):
         return (self.ext_modules and len(self.ext_modules) > 0 or
                 self.rust_extensions and len(self.rust_extensions) > 0)
 
-    Distribution.has_ext_modules = has_ext_modules
+    DistDistribution.has_ext_modules = has_ext_modules
+
+    # this is required because, install directly access distribution's
+    # ext_modules attr to check if dist has ext modules
+    install.orig_finalize_options = install.finalize_options
+
+    def finalize_options(self):
+        ext_modules = self.distribution.ext_modules
+
+        # all ext modules
+        mods = []
+        if self.distribution.ext_modules:
+            mods.extend(self.distribution.ext_modules)
+        if self.distribution.rust_extensions:
+            mods.extend(self.distribution.rust_extensions)
+
+        self.distribution.ext_modules = mods
+
+        self.orig_finalize_options()
+
+        # restore ext_modules
+        self.distribution.ext_modules = ext_modules
+
+    install.finalize_options = finalize_options