+++ setuptools_rust/__init__.py
@@ -14,7 +14,7 @@ import semantic_version
 from . import patch  # noqa
 from .build_ext import build_ext
 
-__all__ = ('RustExtension', 'build_ext', 'build_rust')
+__all__ = ('RustExtension', 'clean_rust', 'build_ext', 'build_rust')
 
 patch.monkey_patch_dist(build_ext)
 
@@ -232,3 +232,34 @@ class build_rust(Command):
                         version, ext.rust_version))
 
             self.build_extension(ext)
+
+
+class clean_rust(Command):
+    """ Clean rust extensions. """
+
+    description = "clean rust extensions (compile/link to build directory)"
+
+    def initialize_options(self):
+        self.extensions = ()
+        self.inplace = False
+
+    def finalize_options(self):
+        self.extensions = [ext for ext in self.distribution.rust_extensions
+                           if isinstance(ext, RustExtension)]
+
+    def run(self):
+        if not self.extensions:
+            return
+
+        for ext in self.extensions:
+            # build cargo command
+            args = (["cargo", "clean", "--manifest-path", ext.path])
+
+            if not ext.quiet:
+                print(" ".join(args), file=sys.stderr)
+
+            # Execute cargo command
+            try:
+                subprocess.check_output(args)
+            except:
+                pass